Split default KMS policy from SNS/SQS-specific one
By default, KMS keys should be accessible by the AWS account's root
user: https://docs.aws.amazon.com/kms/latest/developerguide/key-policies.html#key-policy-default-allow-root-enable-iam

Before, our default KMS policy included extra permissions, like access
from SNS topics and SQS queues. This is necessary in the TDR environments,
but was confusing on mgmt because it doesn't have the message systems
specified in that policy. So the KMS key policies used by Jenkins and
Grafana had extra, confusing permissions.

This change splits the policy into two: one which just grants basic
access to the root user, and one with the file check message systems.

For now, the default is still the policy with the message systems so
that we can switch the environment terraform to the message-specific
policy without making a breaking change in the meantime. A later commit
will switch the default to the basic policy.