Add two file check queue policies.
From the pen test, we had a low priority finding to say that the access
policies for the queues were too permissive. I've updated the the three
file check queues to only allow the download files role to send messages
to it and I've updated the failure queue to only allow the specific
other queues to call it.