Make lambda timeout configurable
This will allow tdr-terraform-environments to set the value, and use
it to determine other values like the corresponding SQS message
visibility timeout.

Set the default to 180 seconds, because this is a common value that
many of the non-file-check Lambdas use.

For now, keep the hard-coded values for Lambdas which do not use the
default values. Once tdr-terraform-environment has been updated to pass
in this parameter, there will be another PR to use the parameter for
all Lambdas.