Add templates for images which will sign commits.
For the plugin updates and npm images, we use to leave Jenkins ECS
plugin to create the tasks directly from the images in ECR.

We can't do that now because as part of the changes to sign commits,
we're adding in environment variables from the parameter store and the
ECS plugin on Jenkins doesn't support that directly. So we need to
create a task definition with the secrets injected and these task
definitions are then references in the Jenkins yml file.

I've also renamed s3publish to sbtwithpostgres as s3publish isn't used
any more and having the name everywhere is confusing.