TDR-1369 Add s3 bucket for cloudfront uploads
I've added in a new bucket for us to set up uploads through Cloudfront
without breaking the config of the existing bucket.

I've also added a new cloudfront module. All it has at the moment is the
origin access identity. This is needed for the S3 bucket as the arn is
added to the bucket policy. I'll add other resources for Cloudfront in a
later PR.