Make team slug optional.
At the moment, we have to approve each one of the parallel jobs because
of the protection rules on the staging and prod environments. This
doesn't make sense for the e2e tests so I've added in an option to not
set a team slug which won't add protection rules.