Add generic lambda module.
We now have to create a few new lambdas and I don't want to have to use
the old system of having a file with a count for each lambda so I've
created this generic lambda module.

Outside of the standard configuration like name and memory size, you can
configure the following:
* SQS queue permissions
* API Gateway lambda invocation permission
* EFS config (Hopefully, we won't need these)
* Plaintext environment variables
* Encrypted environment variables.

The module creates a role with a name that is passed in as a variable.
It also creates a policy from a name and policy string which is passed
in as a variable.

I'm not 100% sure this is the best approach, it might be better to
create the roles and policies outside the lambda module and pass them in
but that will make the root files very verbose so I think this is better
for now. It will be an easy enough change if we need to.