Remove TDR specific templates
The templates for the API gateway definitions and the associated
policies are stored within terraform modules. They should be stored in
terraform environments.

Also, the aws_api_gateway_account only needs to be created once. At the
moment, this is done by checking if the name is ExportAPI which is very
TDR specific and not useful for tdr-terraform-modules.

There's another PR pending for the changes to
tdr-terraform-environments. This doesn't have to be a breaking change
but it will need resources moving in the state using `terraform state mv`