Add https policy to buckets
The pen testers have found a few S3 buckets which didn't have a policy
statement set to deny non https requests.

I've added that statement into each of the affected policies. Once this
is merged, we'll need to update tdr-aws-accounts and
tdr-terraform-environments to the latest version of terraform-modules
and run it for each environment.