Remove the default_statement
There is a statement in this policy which allows TRE to send messages to
this queue but the other statement isn't needed. SQS queues are owned by
default by the account so it should be able to invoke the notifications
lambda without this.

I've tested it and it will still invoke the lambda.

I'll mark this as draft while I check that TRE can still send messages.