Update route53 module.
The module wasn't working when create_hosted_zone was set to false
because it was seting variables as empty string. I've put in a data
block which finds the existing hosted zone and use that for the outputs
and this is working for tdr-aws-accounts. It might have broken something
else.